requirements.txt
PyQt6>=6.5
torch>=2.0
torchvision>=0.15
transformers>=4.30
opencv-python>=4.8
ffmpeg-python>=0.2
Pillow>=10.0
numpy>=1.24
einops>=0.7
kornia>=0.7
timm>=0.9
psutil>=5.9
pytest>=7.0
